fix(js): move normalizeAssets to separate file to avoid heavy imports
The copy-assets plugin imported from copy-assets-handler.ts which
pulls in picomatch, tinyglobby, ignore, and the daemon client. This
caused plugin load failures on CI. Move the pure path functions to
normalize-assets.ts with no external dependencies.
